dockerAirflow/requirements.txt
airflow-clickhouse-plugin==1.1.0
clickhouse_connect
clickhouse-driver
airflow-providers-clickhouse 
openpyxl==3.1.2